requirements.txt
mysql-connector==2.2.9
opencv-python==4.2.0.32
